refactor: replace events collector by get
