+++ pkg/runner/kubectl/describe.go
@@ -1,12 +1,13 @@
-package collect
+package kubectl
 
 import (
 	"errors"
+	"fmt"
 
 	"github.com/kyverno/chainsaw/pkg/apis/v1alpha1"
 )
 
-func Get(collector *v1alpha1.Get) (*v1alpha1.Command, error) {
+func Describe(collector *v1alpha1.Describe) (*v1alpha1.Command, error) {
 	if collector == nil {
 		return nil, nil
 	}
@@ -18,7 +19,7 @@ func Get(collector *v1alpha1.Get) (*v1alpha1.Command, error) {
 	}
 	cmd := v1alpha1.Command{
 		Entrypoint: "kubectl",
-		Args:       []string{"get", collector.Resource},
+		Args:       []string{"describe", collector.Resource},
 	}
 	if collector.Name != "" {
 		cmd.Args = append(cmd.Args, collector.Name)
@@ -32,5 +33,8 @@ func Get(collector *v1alpha1.Get) (*v1alpha1.Command, error) {
 		namespace = "$NAMESPACE"
 	}
 	cmd.Args = append(cmd.Args, "-n", namespace)
+	if collector.ShowEvents != nil {
+		cmd.Args = append(cmd.Args, fmt.Sprintf("--show-events=%t", *collector.ShowEvents))
+	}
 	return &cmd, nil
 }
